Initial Version of Yelp Camp (basics laid out)
App is built using express and ejs. The libraries are required within the app.js file. The campgrounds file is the grid layout for the campgrounds. The landing file is the landing page for initial loading. The new page is a for designed to upload new campgrounds to the app. Both partial files are html code named for the corresponding part of the pages. These files have been included accross multiple pages.